v2023.8.0-beta15 fixes #163, 162, 161

+++ custom_components/smart_irrigation/frontend/src/views/zones/view-zones.ts
@@ -394,7 +394,7 @@ class SmartIrrigationViewZones extends SubscribeMixin(LitElement) {
                 @change="${(e: Event) =>
                   this.handleEditZone(index, {
                     ...zone,
-                    [ZONE_MODULE]: parseInt(
+                    [ZONE_MODULE]: parseFloat(
                       (e.target as HTMLSelectElement).value
                     ),
                   })}"
